+++ src/tools/index.js
@@ -1,1 +1,4 @@
 export * from './request';
+export * from './cookies';
+export * from './object';
+export * from './alertSnackbar';