+++ src/views/ValidateCredentials/index.js
@@ -1,33 +1,46 @@
-import React, { useState } from 'react';
-import { useSelector } from 'react-redux';
+import React, { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 
-import Login from './components/Login';
+// import Login from './components/Login';
 import Register from './components/Register';
-// import Unauthorized from './components/Unauthorized';
 import { Error401 } from 'views';
+import {
+  getCookie,
+  getObjectFromString,
+  USER_SUCCESSFULLY_LOADED,
+  USER_KEY,
+} from 'tools';
+import { showAlertSnackbar, storeUserFoundOnCookies } from 'ducks';
 
 const ValidateCredentials = () => {
-  const [status] = useState('register');
+  // const [status] = useState('register');
+  const dispatch = useDispatch();
   const { data } = useSelector((state) => state.user);
+
+  useEffect(() => {
+    const dataFromCookies = getObjectFromString(getCookie(USER_KEY));
+
+    // if data is null set the info from cookies
+    if (!data) {
+      if (!dataFromCookies) {
+        const noValues = {
+          searchParams: null,
+          data: null,
+        };
+
+        dispatch(storeUserFoundOnCookies(noValues));
+      } else {
+        dispatch(storeUserFoundOnCookies(dataFromCookies));
+        dispatch(showAlertSnackbar(USER_SUCCESSFULLY_LOADED));
+      }
+    }
+  }, [data, dispatch]);
+
+  // if data is null return error401, if not just return register
+  return !data ? <Error401 /> : <Register />;
 
   // TODO: check if the credentials are valid and if
   // will register or login
-
-  // TODO: discutir si tiene sentido la vista de que no figura
-  // en el padron, ya que no tiene sentido porque de eso depende
-  // la base de datos y si no esta en el padron ni si quiera aparecera
-  // un resultado al hacer una request a la api
-
-  // TODO: discutir el tema de la contrasena y si seria mejor que
-  // me envie un status (booleano) de registrado o no para evitar tener que
-  // recibir en el body el password hasheado
-
-  if (!data) return <Error401 />;
-
-  if (status === 'login') return <Login />;
-  // if (status === 'register') return <Register />;
-
-  return <Register />;
 };
 
 export default ValidateCredentials;